refactor common functionality between task cell and schedule cell into a parent class
