+++ src/main/java/com/ateam/onpoint/gui/components/ScheduleCell.java
@@ -1,0 +1,4 @@
+package com.ateam.onpoint.gui.components;
+
+public class ScheduleCell extends Cell {
+}
