+++ src/main/java/com/ateam/onpoint/gui/components/TaskCell.java
@@ -18,13 +18,10 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 
-public class TaskCell extends ListCell<Task> {
-    private final HBox root = new HBox();
+public class TaskCell extends Cell {
     private final Label description = new Label();
-    private final CheckBox checkBox = new CheckBox();
     private final Label dateLabel = new Label("yyyy/mm/dd hh:mm");
     private final Label durationLabel = new Label("00h 00m");
-    private final ContextMenu contextMenu;
 
     public TaskCell() {
         super();
@@ -49,8 +46,6 @@ public class TaskCell extends ListCell<Task> {
 
         this.description.setWrapText(true);
 
-        this.contextMenu = createContextMenu();
-
         this.root.setAlignment(Pos.CENTER_LEFT);
         this.root.prefHeight(15);
         this.root.prefWidth(450);
@@ -58,8 +53,6 @@ public class TaskCell extends ListCell<Task> {
         this.root.setOnContextMenuRequested(e -> this.contextMenu.show(this.root.getScene().getWindow(), e.getScreenX(), e.getScreenY()));
 
         this.dateLabel.setPadding(new Insets(0, 5, 0,0));
-
-        setupDragAndDrop();
 
         this.root.getChildren().addAll(
                 this.checkBox,
@@ -73,90 +66,6 @@ public class TaskCell extends ListCell<Task> {
                 new Spacer(5),
                 this.dateLabel
         );
-    }
-
-    private void setupDragAndDrop() {
-        this.root.setOnDragDetected(e -> {
-            if (getItem() == null) {
-                return;
-            }
-
-            final var dragBoard = startDragAndDrop(TransferMode.MOVE);
-            final var clipBoard = new ClipboardContent();
-            clipBoard.putString(this.getItem().descriptionProperty().get());
-            dragBoard.setDragView(this.root.snapshot(new SnapshotParameters(), null));
-            dragBoard.setContent(clipBoard);
-
-            e.consume();
-        });
-
-        this.root.setOnDragOver(e -> {
-            if (e.getGestureSource() != this && e.getDragboard().hasString()) {
-                e.acceptTransferModes(TransferMode.MOVE);
-            }
-            e.consume();
-        });
-
-        this.root.setOnDragEntered(e -> {
-            if (e.getGestureSource() != this && e.getDragboard().hasString()) {
-                this.setOpacity(0.3);
-            }
-        });
-
-        this.root.setOnDragExited(e -> {
-            if (e.getGestureSource() != this && e.getDragboard().hasString()) {
-                this.setOpacity(1);
-            }
-        });
-
-        this.root.setOnDragDropped(e -> {
-            if (this.getItem() == null) {
-                return;
-            }
-
-            final var dragboard = e.getDragboard();
-            var success = false;
-
-            if (dragboard.hasString()) {
-                final var items = getListView().getItems();
-                final var draggedItem = items.stream()
-                        .filter(c -> c.descriptionProperty().get().compareTo(dragboard.getString()) == 0)
-                        .findFirst()
-                        .orElse(null);
-                Objects.requireNonNull(draggedItem);
-
-                final var draggedItemIndex = items.indexOf(draggedItem);
-                final var thisIndex = items.indexOf(getItem());
-
-                items.set(draggedItemIndex, getItem());
-                items.set(thisIndex, draggedItem);
-
-                List<Task> itemsCopy = new ArrayList<>(getListView().getItems());
-                getListView().getItems().setAll(itemsCopy);
-
-                success = true;
-            }
-
-            e.setDropCompleted(success);
-            e.consume();
-        });
-
-        setOnDragDone(DragEvent::consume);
-    }
-
-    private @NotNull ContextMenu createContextMenu() {
-        final var contextMenu = new ContextMenu();
-
-        final var editTaskMenuItem = new MenuItem("Edit Task");
-        editTaskMenuItem.setOnAction(e -> {
-            TaskEditor.getInstance().openTaskEditor(getScene().getWindow(), this.getItem());
-        });
-
-        final var deleteTaskMenuItem = new MenuItem("Delete");
-        deleteTaskMenuItem.setOnAction(e -> TaskDatabase.getInstance().removeTask(this.getItem()));
-
-        contextMenu.getItems().addAll(editTaskMenuItem, deleteTaskMenuItem);
-        return contextMenu;
     }
 
     @Override
